example/requirements.txt
django-oscar[sorl-thumbnail]
pinax-announcements
django-background-tasks
django-tables2
treebeard
django-haystack
-e ..
